SDK-721: Make CreateExceptionFromStatusCode generic for reuse, move exceptions to own folder
