+++ src/Yoti.Auth/Exceptions/AmlException.cs
@@ -1,0 +1,15 @@
+﻿using System;
+
+namespace Yoti.Auth.Exceptions
+{
+    public class AmlException : Exception
+    {
+        public AmlException(string message) : base(message)
+        {
+        }
+
+        public AmlException(string message, Exception innerException) : base(message, innerException)
+        {
+        }
+    }
+}
