--- src/Yoti.Auth/YotiProfileException.cs
@@ -1,15 +1,0 @@
-﻿using System;
-
-namespace Yoti.Auth
-{
-    public class YotiProfileException : Exception
-    {
-        internal YotiProfileException(string message) : base(message)
-        {
-        }
-
-        internal YotiProfileException(string message, Exception innerException) : base(message, innerException)
-        {
-        }
-    }
-}
